Adding GET /users/:id and some refactor
Added route for users/:id to get a single user. Moved all
user helpers into a new file under the utils/ directory.
