+++ routes/classes.js
@@ -1,5 +1,5 @@
 const express = require('express');
-const {  getAllClasses } = require('./databaseAccess');
+const {  getAllClasses } = require('../db');
 const router = express.Router();
 
 router.get("/", async function(req, res, next){
